+++ server/src/models/Interaction.ts
